+++ src/confcom/HISTORY.rst
@@ -7,6 +7,7 @@ Release History
 +++++
 
 * Added de-duping of exec_processes in the generated policies so same command used in two probes results in one exec_processes entry
+* Add MacOS support for Linux-container policy generation
 
 2.1.0
 +++++